pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tpex"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2"
]

[tool.setuptools]
package-dir = {""="src"}

[tool.setuptools.packages.find]
where = ["src"]